db_saver/pyproject.toml
[tool.poetry]
name = "db-saver"
version = "0.1.0"
description = ""
authors = ["Максим Жданов <exodus.yarwood@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.32"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.8"
alembic = "^1.13.2"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
